requirements.txt
yfinance==0.2.44
pandas==2.2.3
google-cloud-storage==2.18.2
google-cloud-logging==3.11.3
functions-framework==3.*